Data: Update script to download the wiki page
