+++ package.json
@@ -15,6 +15,7 @@
     "test": "vitest run",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage",
+    "data:wiki:downloadWiki": "bun ./data/wiki/downloadWiki/downloadWiki.ts && prettier ./data/ --write",
     "updatePackages:major": "npx taze major --includeLocked --write --exclude eslint && npm run updatePackages:minor && npm install",
     "updatePackages:minor": "npx taze minor --includeLocked --write && npm install",
     "prepare": "husky"
@@ -37,6 +38,7 @@
     "@tailwindcss/typography": "^0.5.15",
     "@testing-library/react": "^16.0.1",
     "@total-typescript/ts-reset": "^0.6.1",
+    "@types/bun": "^1.1.9",
     "@types/eslint": "^8.56.12",
     "@types/node": "^20.16.5",
     "@types/react": "^18.3.5",
